Major rewrite of CaseStudy implementation
Split CaseStudy into two parts, BasicCaseStudy and AssetCaseStudy.
BasicCaseStudy is just the first page of data, with no assets attached.
AssetCaseStudy adds the image and press assets to the object. Added
an empty implementation of DownloadableAsset to use when saving
CaseStudy initially.

+++ src/main/scala/wknyc/model/Content.scala
@@ -45,7 +45,7 @@ object Content {
 }
 /** Trait to be mixed into objects which must maintain a specific orderering */
 trait Ordered {
-	def position:Int
+	def position:Long
 }
 object Ordered {
 	val NodeType = "wk:ordered"
